Updated the plotting of spectra
Added spectralon normalisation button to display the diffuse reflection spectrum normalised by a spectralon spectum.

Added option to normalize the spectrum with either the currently loaded spectralon file or the spectralon spectrum saved in the  currently displayed acquisition file.

Fixed the legend that was not updating when a set of new lines are plot